feat: move error.rs file
